+++ LiVHY/src/util/types.h
@@ -6,9 +6,9 @@
 #include "glm/vec2.hpp"
 
 
-using RGB = glm::dvec3;
-using RGBA = glm::dvec4;
-using Vec3 = glm::dvec3;
+using RGB = glm::vec3;
+using RGBA = glm::vec4;
+using Vec3 = glm::vec3;
 using Vec3i = glm::ivec3;
-using Vec2 = glm::dvec2;
+using Vec2 = glm::vec2;
 using Vec2i = glm::ivec2;
